server/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
atlassian-python-api==4.0.7
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Deprecated==1.2.18
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
jmespath==1.0.1
oauthlib==3.3.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
requests-oauthlib==2.0.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3